Remove RSpec Rake task
When trying to run `heroku run rake db:create`, Heroku logged an error
that it couldn't find the RSpec rake task. My assumption is that RSpec
is a development/test dependency, so it probably isn't bundled on the
Heroku server.

As a result, this commit removes the Rake task and updates the Travis
config to run `rspec` instead of `rake` as the test command.